requirements.txt
flask
flask-cors
google-generativeai
pypdf2
python-dotenv
